refactor: Point 클래스 제거 및 로직 이동
