+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -8,10 +8,15 @@ import static blackjack.domain.Fixtures.TEN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.cards.Cards;
+import blackjack.domain.cards.card.Card;
+import blackjack.domain.cards.card.denomination.Denomination;
+import blackjack.domain.cards.card.denomination.Suit;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 public class CardsTest {
     private Cards cards;
@@ -59,23 +64,29 @@ public class CardsTest {
     }
 
     @Test
-    @DisplayName("단순 포인트 리턴 기능 테스트")
-    void getRawPointTest() {
-        assertThat(cards.getRawPoint())
-                .isEqualTo(22);
-    }
-
-    @Test
-    @DisplayName("특정 카드 개수 리턴 기능 테스트")
-    void getAceCountTest() {
-        assertThat(cards.getAceCount())
-                .isEqualTo(2);
-    }
-
-    @Test
     @DisplayName("카드 리스트 리턴 기능 테스트")
     void get() {
         assertThat(cards.getCopy())
                 .isEqualTo(List.of(ACE, ACE));
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {
+            "ACE,FIVE,ACE,17",
+            "JACK,FIVE,JACK,25",
+            "ACE,ACE,ACE,13",
+            "JACK,QUEEN,KING,30",
+            "ACE,TEN,TEN,21",
+    })
+
+    @DisplayName("올바른 포인트 계산되는지 테스트")
+    void setExpectedPointTest(Denomination d1, Denomination d2, Denomination d3, int expectedPoint) {
+        Cards cards = new Cards();
+        cards.add(Card.of(d1, Suit.SPADE));
+        cards.add(Card.of(d2, Suit.SPADE));
+        cards.add(Card.of(d3, Suit.SPADE));
+
+        assertThat(cards.getPoint())
+                .isEqualTo(expectedPoint);
     }
 }
